Update API to support returning Option and Answer together.
Previously the API had resolvers for createOtherAnswer and deleteOtherAnswer.

These have been renamed to createOther and deleteOther respectively.
The latest graphql schema changes return both the other Option and Answer as a pair
from the mutation.

This was required to allow optimistic updating of the Apollo cache when adding/removing
the other option for checkbox/radio answers.

I have also updated a lot of the tests to reflect this change.

Update yarn.lock