Fix null description for other answer.
There is an issue where null values are being returned for other answer description.

This change to the API should ensure that other answers have an empty string
description by default.

This is only an issue because the other answer is a hidden field. So users cannot
provide a value.