Duplication of pages
Notes:

1. Jest not closing tests as handles left open
This was caused by a failing rollback migration causing the db handle
to be left open and so jest/node was not able to finish. The migration
rollback is 20180622154505_migrate_mandatory_field_data_to_properties

2. Remove functions we aren't using and add tests to functions we do use
We cannot remove the resolvers as it technically would be a breaking
api change.

3. Change options ordering. It was previously ordered by created at but
this was all the same when duplicating so we have changed it to use id
instead.