Duplicate performance
This changes the way the duplication works.

his is an attempt to remove a lot of the complexity around how
duplication was implemented and also change the architecture to improve
speed.

This is done by instead of working through the tree entity by entity
(and waiting for each entity to be completed) we now work top down
through the hierarchy so we create all parents before we create any
children. We also parallelise the creation of parents when they share no
relationships. All of this is done in the `duplicateTree` function in
`duplicateStrategy/utils.js`. It takes a defined hierarchy and works
through.

Our structure is nearly perfectly compatible with this way of working
excluding piping (where pages reference answers) and routingDestinations
which themselves have no reference to the parent. So in both these cases
we have to query those objects again to work out what to update or
duplicate.

UKIS duplicate speed change from ~4s to ~400ms